+++ src/index.js
@@ -5,8 +5,6 @@ import reportWebVitals from './reportWebVitals'
 import 'bootstrap/dist/css/bootstrap.css'
 import App from './app/App'
 
-//import Users from './app/components/users'
-
 const root = ReactDOM.createRoot(document.getElementById('root'))
 root.render(
   <React.StrictMode>
